refactor: update files and folders name
