+++ src/features/register/services/registerCompany.ts
@@ -1,5 +1,5 @@
 import { httpClientFetch } from '@m-care/features/@shared/lib'
-import { RegisterCompanyFormData } from '../components/RegisterCompanyForm'
+import { RegisterCompanyFormData } from '../components/register-company-form'
 
 export const registerCompany = async (data: RegisterCompanyFormData) => {
   const [error, response] = await httpClientFetch({
